+++ zerver/openapi/python_examples.py
@@ -1367,6 +1367,30 @@ def get_stream_topics(client: Client, stream_id: int) -> None:
     validate_against_openapi_schema(result, "/users/me/{stream_id}/topics", "get", "200")
 
 
+@openapi_test_function("/users/me/apns_device_token:post")
+def add_apns_token(client: Client) -> None:
+    # {code_example|start}
+    request = {"token": "apple-tokenbb", "appid": "org.zulip.Zulip"}
+    result = client.call_endpoint(url="/users/me/apns_device_token", method="POST", request=request)
+    # {code_example|end}
+
+    validate_against_openapi_schema(result, "/users/me/apns_device_token", "post", "200")
+
+
+@openapi_test_function("/users/me/apns_device_token:delete")
+def remove_apns_token(client: Client) -> None:
+    # {code_example|start}
+    request = {
+        "token": "apple-tokenbb",
+    }
+    result = client.call_endpoint(
+        url="/users/me/apns_device_token", method="DELETE", request=request
+    )
+    # {code_example|end}
+
+    validate_against_openapi_schema(result, "/users/me/apns_device_token", "delete", "200")
+
+
 @openapi_test_function("/typing:post")
 def set_typing_status(client: Client) -> None:
     ensure_users([10, 11], ["hamlet", "iago"])
@@ -1656,6 +1680,8 @@ def test_users(client: Client, owner_client: Client) -> None:
     get_alert_words(client)
     add_alert_words(client)
     remove_alert_words(client)
+    add_apns_token(client)
+    remove_apns_token(client)
 
 
 def test_streams(client: Client, nonadmin_client: Client) -> None: